LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-focus-within-expected.txt



FAIL CSS Overflow Test: focused ::scroll-marker sets :focus-within on its ::scroll-marker-group, scroller and all the ancestors assert_equals: focused ::scroll-marker sets :focus-within on its ::scroll-marker-group expected "rgb(0, 128, 0)" but got ""
